requirements.txt
requests
line-bot-sdk
flask
pyngrok
google-api-python-client
oauth2client
gspread
google-generativeai
pillow